fix: Move DrawerManager comonponent in common folder
